🤖 Merge 8.6.x-sync into master

+++ cylc/flow/remote.py
@@ -79,16 +79,18 @@ def get_proc_ancestors():
 def watch_and_kill(proc):
     """Kill proc if my PPID (etc.) changed - e.g. ssh connection dropped."""
     gpa = get_proc_ancestors()
-    # Allow customising the interval to allow tests to run faster:
-    interval = float(os.getenv('CYLC_PROC_POLL_INTERVAL', 60))
+    interval = 1  # secs
+    count = 0
     while True:
-        sleep(interval)
+        count += 1
         if proc.poll() is not None:
             break
-        if get_proc_ancestors() != gpa:
+        if round(count * interval) % 60 and get_proc_ancestors() != gpa:
+            # (Only run ps command once a minute)
             sleep(1)
             os.kill(proc.pid, signal.SIGTERM)
             break
+        sleep(interval)
 
 
 def run_cmd(
